Implement Loading and Saving to a file
This change implements the following 3 file-related operations:
- Creating a new document. This prompts the user for confirmation
and if approved clears all previously generated HTML to start over.
- Saving the generated HTML to a file. This opens a dialog where
the user can browse the local file system and choose the location
for saving the generated HTML. Opens in $USER_HOME by default.
- Loading HTML from already existing file. Again, opens a dialog
for browsing. When the user chooses a file, the file is parsed
and validated against an XSD schema that describes what a valid
HTML document is. If the HTML document is invalid, it shows a
warning and does not load it.