Partial implementation of drag & drop functionality
This change creates two panes in the designer view.
The right-hand side one holds all available HTML elements.
These elements can be dragged & dropped into the left-hand side pane.
When the user decides to generate the HTML (should happen with the
click of a button), all elements that are inside the left-hand
side pane will get translated to HTML.