Improve progress loss validation
Previously, the validation that checks if the user will lose their
progress by creating a new file or opening an exisiting one only
checked if the preview text area is empty. Now, it also checks
if the designer pane is empty, too. If any of the two is not empty,
the warning dialog is displayed.