Implement HTML generation
This change adds the quintessential part of this application - the actual
generation of HTML from GUI objects. This is achieved by finding all
objects of type DomElement that have been dragged to the Designer pane and
calling the toHtml method on each of them. The order in which the elements
appear in the generated HTML is top to bottom, meaning the top-most
elements appear first in the HTML document.

Also added two new supported dom elements - subscript and superscript.

Implemented the About dialog, too.